Energy Balance Equation added, with first attempt at FluxBC for it.
Testing concluded that this configuration (product of n_e and mean 
electron energy) produces a negative result for the product, which is 
obviously not realistic. I think this is due to an incorrect combination 
of the mean electron energy boundary condition given in the paper and 
energy flux equation in order to make an energy flux BC. I need to 
break-up this product in the eventual implementation for a second 
attempt.