Simplify and generalize energy MMS test input file and python script.
Still having discrepancies, though they are now smaller....Eliminating 
kernels and re-evaulating the RHS has narrowed down the potential 
trouble spots to the grad(energy_flux) term. Though the implemented code 
is identical to what has been derived (and re-derived). There may yet be 
a bug in the python script, but generalizing the calculation a bit 
doesn't seem to make much of a difference.